do not upload mesh data onto GPU if already there

+++ src/Engine/shaders/animatedModel.vert
@@ -1,11 +1,11 @@
 #version 450
 #extension GL_ARB_separate_shader_objects : enable
 
-layout (location = 0) in vec3   position;
-layout (location = 1) in vec3   normal;
+layout (location = 0) in vec3    position;
+layout (location = 1) in vec3    normal;
 layout (location = 2) in uvec4   boneIdx;
-layout (location = 3) in uvec4   boneWeights;
-layout (location = 4) in vec2  uv;
+layout (location = 3) in vec4    boneWeights;
+layout (location = 4) in vec2    uv;
 
 layout (location = 0) out vec2  out_uv;
 layout (location = 1) out vec3  out_normal;
